simulation of a propagating sinusoidal striking a dielectric medium

+++ fd1d/program/fd1d_1_4.py
@@ -13,7 +13,7 @@ plt.style.use("classic")
 plt.style.use("../pyplot.mplstyle")
 
 
-def visualize(ns: int, nx: int, epsr: int, cb: np.ndarray, ex: np.ndarray, hy: np.ndarray) -> None:
+def visualize(ns: int, nx: int, epsr: float, cb: np.ndarray, ex: np.ndarray, hy: np.ndarray) -> None:
 	fig, (ax1, ax2) = plt.subplots(2, sharex=False, gridspec_kw={"hspace": 0.2})
 	fig.suptitle(r"FDTD simulation of a sinusoidal wave striking dielectric material")
 	medium = (0.5/cb - 1)/(epsr - 1) if epsr > 1 else (0.5/cb - 1)
